core: VM Numa command test setup unification
Reworked the VM numa command test classes to have a unified setup,
as they all currently copy-paste very similar boiler-plate code.
As part of this reworking, a single command instance is annotated with
a @Spy annotation in order to prepare this test for injecting DAOs as
done in the subsequent patches.

With this patch, the class is left with a single command instance
which has a single instance of a list of numa nodes.
Since these issues remain between tests, they need to be cleaned up in
an @After method.

Change-Id: Id6e9746f4cba3767e4948ae066149226927ee352
Signed-off-by: Allon Mureinik <amureini@redhat.com>
Signed-off-by: Tal Nisan <tnisan@redhat.com>