restapi: Separate DNS server addresses with spaces
Currently when a set of DNS servers is provided as the initialization
data for a virtual machine the result is concatenated and saved to the
database, without spaces. For example, if the following data passed to
the API:

  <vm>
    <initialization>
      <cloud_init>
        <network_configuration>
          <dns>
            <servers>
              <host>
                <address>192.168.0.1</address>
              </host>
              <host>
                <address>192.168.0.2</address>
              </host>
              <host>
                <address>192.168.0.3</address>
              </host>
            </servers>
          </dns>
          ...
        </network_configuration>
        ...
      </cloud_init>
    </initialization>
    ...
 </vm>

Then the following is stored in the database (note that there are no
spaces separating the addresses):

  192.168.0.1192.168.0.2192.168.0.3

And the following is added to the generated cloud-init configuration
file:

  dns-nameservers 192.168.0.1192.168.0.2192.168.0.3

This way to provide the cloud-init configuration is deprecated, the
preferred way is to use directly the 'initialization' element:

  <vm>
    <initialization>
      <dns_servers>192.168.0.1 192.168.0.2 192.168.0.3</dns_servers>
      ...
    </initialization>
    ...
  </vm>

But we should fix the issue anyhow. That is what this patch does.

Change-Id: I417bd7b1da69bda0d3b2879ee519ea4744fd35a9
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>