core: Updating RNG devices on Cluster change
urandom/random RNG device of cluster-aware VM-like entities needs to be
updated when compatibility version of associated cluster is changed
because we support 'random' RNG source for version 4.0 and lower and
'urandom' RNG source for 4.1 and higher.

Custom compatibility version is honored allowing to run VM in 4.0
cluster with `urandom` device and VM in 4.1 cluster with `random`.
However engine doesn't ensure that these devices will be present on
hosts. Thus using custom compatibility level to get different RNG device
is considered unsafe and should only be used to preview RNG generator
behavior in non-critical VMs.

This patch ensures that whenever cluster version is changed or cluster
associated with VM or template is changes then RNG device of related
entities are appropriately updated.

Change-Id: I4048c2ce32c63455973f82da4175a7ec2a12a76c
Bug-Url: https://bugzilla.redhat.com/1374227
Signed-off-by: Jakub Niedermertl <jniederm@redhat.com>