webadmin,userportal: Fix JqueryUtils.getTextFromHtml
oVirt GWT UI currently uses jQuery version 1.11.3

Quote from jQuery upgrade guide [1]:

"
As of 1.9, a string is only considered to be HTML if it starts
with a less-than ("<") character.
..
If a string is known to be HTML but may start with arbitrary text
that is not an HTML tag, pass it to jQuery.parseHTML() which will
return an array of DOM nodes representing the markup.
"

[1] http://stage.jquery.com/upgrade-guide/1.9/#jquery-htmlstring-versus-jquery-selectorstring

The problem was that jQuery treated `html` argument as selector
if the string didn't start with `<` character. Because of this,
the string failed to parse as selector, resulting in error like:

  Syntax error, unrecognized expression: Dev. Size

This patch fixes the problem.

Change-Id: Ic939c09f3cb2e2a80b6ef2dd15cb759cbc27f946
Bug-Url: https://bugzilla.redhat.com/1386765
Signed-off-by: Vojtech Szocs <vszocs@redhat.com>
(cherry picked from commit fc09bce61ea505c67399a6c678c329de456257f7)