engine : sdm create VM from template
When creating a VM from template as Clone/Independent,
the template disks are copied.
This operation can consume a lot of time and resources.

Until now data operations on the hosts were performed only on the elected
SPM, causing a potential bottleneck.

In this patch, the copy operation of the disks can be performed also on
other hosts, not only the SPM.

Change-Id: I9358478d9f76804ea6a71c2a8895dcc512e6b52c
Signed-off-by: Fred Rolland <frolland@redhat.com>