core: remove vm usage related histories from the database
Now that VmStatistics are cached in VmManager, there is no need
to dump several things they contain to the database. However,
since we currently focus on easy and non-risky things, this patch
addresses specifically the memory/cpu/network usage histories -
these are really redundant being the database as they are only
used by the webadmin that gathers the information using a
(search) query that takes this information from VmManager.

Unfortunately, we can neither drop the whole statistics from the
database nor reduce the frequency of the updates at this stage.
This kind of changes can improve performance on large-scale setups
significantely since they reduce the amount of data dump into the
database substantially.

Change-Id: If72063205907c21d9bd80610a1c1ea1356a9c5e6
Signed-off-by: Arik Hadas <ahadas@redhat.com>