core: remove unneeded values of vm statistics from the database
Saving VM statistics in a database is somewhat problematic since
it is expected to change on every statistics cycle and in large-scale
setups it leads to high amount of data being written to the database
very frequently. Unfortunately, we cannot eliminate it completely
since DWH and the dashboard get their information by querying the
database directly.

However, we can reduce the amount of data that is being written to
the database by removing the fields that no one queries (but are
taken from the cached VM statistics instead). There are currently
four such fields that are removed in this patch. Together with
If7206320, more than 50% of the fields that were written to
vm_statistics in 4.0 are no longer written to the database.

Change-Id: I1b7edbfebbb3bb83d1c1c6f08144dda269c83729
Signed-off-by: Arik Hadas <ahadas@redhat.com>