core: Add IConfigUtilsInterface#refresh()
Move the refreshing logic to the interface so it can be called
generically, without having to depend on the DBConfigUtils specific
implementation.

This is required in order to change
ConfigUtilsBase#getValue(VdcOption) to an instance method, which allows
it to call the polymorphic getValue(ConfigValues, String), thus solving
the regression introduced in 32a5e8, where value dependent config
values retrieved ONLY the defaults even if a value was present in the
database.

Change-Id: I0b0358ae542fdf1d6b9265ce7cdca8adcaf91d83
Signed-off-by: Allon Mureinik <amureini@redhat.com>