engine : User is not able to edit profile
User who login using group permissions
does not exist in the database as of 4.0
SSO rework.

This is causing issues with editing user
profiles which requires user to exists in
the database and also with other operations
where the user inherits the permissions of
the group.

Reintroducing saving users to database
resolves these issues.

The user info is also updated in users table on
user login if the information has changed.

Change-Id: Ie83bc67585a6af7276289e7898852e998f68671d
Bug-Url: https://bugzilla.redhat.com/1389251
Signed-off-by: Ravi Nori <rnori@redhat.com>