core: do not instantiate vm-manager when not needed
Now that the instantiation of VmManager involves a query for
the static data of a VM from the database, we should be more
careful not to make redundant instantiations.

This patch adds a machanism to retrieve the manager object
of a specific VM only if it is already instantiated. This
mechanism is used by methods that query content that is
typically contained in VmManager, but when no such object
exists, should use default values.

Change-Id: I18e5ed3da7f1d5ee1b2a987f08eec4ba8965919a
Bug-Url: https://bugzilla.redhat.com/1392209
Signed-off-by: Arik Hadas <ahadas@redhat.com>