core: remove vm guest agent interfaces in a single query
The monitoring aggregates the changed VM guest agent interfaces
but then iterates over all the VMs with changed interfaces and
remove their interfaces one-by-one. Interacting with the DB
within loop has never been a best-practice.

This patch changes the removal of VM guest agent interfaces so
they will be removed for all the given VMs in a single query.

Note that while a better solution would be to have some kind
of identifier for raws in vm_guest_agent_interfaces (and then
we won't need to remove and add all the interfaces on every
change) seems possible, this is a less-risky change that can
be backported.

Change-Id: I89e89a116b9247171e7708704b14c2b8c7b41943
Signed-off-by: Arik Hadas <ahadas@redhat.com>