core: StorageJobCallback - unreported job when entity polling isn't supported
This patch adds modifies StorageJobCallback to support handling of commands that
doesn't support entity polling (for example: Storage Domain commands) when the
executed job isn't reported by the performing host. That scenario might happen as job are
cleared randomally from the host report.

When a job isn't reported, we can determine the job status by using entity
polling if it is supported (as the result will be determined by the entity).
If entity polling isn't supported - the callback will now assume the job has
failed so that the command will end. There is no point in continue polling
as the status isn't reflected on the entity and the job status was cleared from
the host report.

Change-Id: I6a79ef2ac1c531c97e2d5dc169c785b5567d773f
Signed-off-by: Liron Aravot <laravot@redhat.com>