core: Move SCSI reservation indication to DiskVmElement
As the indication to whether a disk uses SCSI reservation is relevant to
disks and not all the other types of devices it should be moved from
VmDevice to DiskVmElement.
Aside for that the fact that the LunDisk entity contains that indication
is plain wrong and redundant and thus removed.
Additionally a test was added to MigrateVmCommand which is the most
important command regarding the SCSI reservation

Bug-Url: https://bugzilla.redhat.com/1362152
Change-Id: I3e70748464637720edc2d023becd8d8bcf648ce1
Signed-off-by: Tal Nisan <tnisan@redhat.com>