webadmin,userportal: Clean up Editor fields before Model objects
This ensures that GWT widgets hooked to UiCommon models have
a chance to do their cleanup *before* we destroy the models.

This fixes NPEs where a model is already destroyed but a GWT
widget attempts to access its properties such as the property
change event, see AbstractUiCommandButton#cleanup for example.

Change-Id: I3dd8e320a774c35db8262e9a297557243a99ef5a
Bug-Url: https://bugzilla.redhat.com/1389998
Signed-off-by: Vojtech Szocs <vszocs@redhat.com>