core: Remove connection check before each query
Removing connection validation since it is executed for each connection
loads the database (120,000 calls on 15 min for scale env
with 50 hosts and 1500 VMs, this was the topmost query!!!)
(The load is more on the round-trip than cpu/mem because this query is a
totally inlined and has no planner usage)

The reason for that are:

1) connection failures are rare
2) there is not guarantee that after the "select 1;" query is executed
that the following query will succeed
3) the connection pool generally checks its connections
4) if the connection is not OK, we should rely on exception handling

This patch turns this validation to be in the background and not occur
for each connection that is taken from the pool.
Default time for background validation was set to 5 min.

Change-Id: Ie02de4738f542f399111eddf2272523143a6f135
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1388421
Signed-off-by: emesika <emesika@redhat.com>