engine: Optimize performance of host monitoring
HostMonitoring updates vds dynamic
data every 3 seconds as determined
by the vdc_option VdsRefreshRate.

It is not necessary to update the data
if it has not changed since the last
update.

The equals method of VdsDynamic has been updated
to skip comparing numaNodeList which is not
persisted in the database.

Change-Id: I2fb8e470e12d34af80a0d84d51db2049917a5612
Bug-Url: https://bugzilla.redhat.com/1388536
Signed-off-by: Ravi Nori <rnori@redhat.com>