core: Don't run checkstyle on modules/dependencies
A recent patch changed the checstyle configuration so that it runs in
all the sub-projectes inside the 'backend' directory. This means that it
also runs for the 'backend/modules/dependecies' sub-project. But this
sub-project doesn't contain source code, so running checkstyle doesn't
help, and it actually prevents some usages of this sub-project. For
example, we currently do the following from some CI jobs in order to
analyze the dependencies of the project:

  $ mvn clean package --projects \
  backend/manager/dependencies,\
  backend/manager/dependencies/common,\
  backend/manager/dependencies/tools \
  -DskipTests

Now this fails becaus it can't find the artifact that contains our
custom checstyle rules. To avoid this failure this patch moves the
configuration of the checstyle plugin to the sub-directories of the
'backend' directory that do contain source code, where running checstyle
is useful.

Change-Id: I3630e380b742d100801cb5a1c176ef8a52f3a46b
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>