core: Move/Copy disk flow - weight infra and usage
When command is executed and defines a Job, its initial job weight
is 100. If that command is executing child commands, it can split
the job weight between them to represent the part of each command
in the job.

This patch adds the infra to calculate and pass the weight
of executed child commands and makes use of that infra in the
move/copy disk SDM flow.

In order to avoid handling double values, we use integer for
the weight of each commands - CommandsWeightsHelper.adjust()
can be used in order to round and split the weight of the child
commands according to the provided total weight.

Change-Id: I5414609a17fd06dbc73591487e34619316e71ddb
Signed-off-by: Liron Aravot <laravot@redhat.com>