core: Propagate UpdateVm failure to UpdateClusterCommand
When changing cluster compatibility version with UpdateClusterCommand,
UpdateVmCommand is called for every VM in the cluster. If
UpdateVmCommand fails, failure details are not propagated to
UpdateClusterCommand and the failure is displayed to user as generic
"Internal error".

The patch fixes this issue.

Change-Id: Ib68f691db5027106c582ed6ecd2d53c7d82e3454
Bug-Url: https://bugzilla.redhat.com/1373573
Signed-off-by: Shmuel Melamud <smelamud@redhat.com>
(cherry picked from commit 9a281de9383329ed9a92ec353cf33910f91ea776)