frontend: use authz name instead of profile name for sysprep
VM dialog previously pre-filled value for domain in sysprep tab,
as profile name. This is incorrect behaviour, as profile name
is just user visible name used only for login. Authz name is
the name of the authorization provider, which should reflect
real domain name of the windows machine. This patch changes that.

Change-Id: Ib19f4c4c51b1a9c62ca68afee330b9aec9893b88
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1380128
Signed-off-by: Ondra Machacek <omachace@redhat.com>