webadmin: Fix custom bond mode option input
The label and the input for "Custom Mode" options in the
SetupNetworksBondPopupView was only visible, if the bonding
mode is chosen to 'Custom'. In the previous implementation the higher
level container in the user interface did not provide enough space to
receive user inputs for "Custom Mode" options. Without choosing the
"Custom Mode", the user was not aware of the input of "Custom Mode"
options.

The higher level container in the user interface is increased and
they layout is changed to a Container of bootstrap grid system.
Additionally the "Custom Mode" options input is disabled, if the user
does not choose the "Custom Mode".

Change-Id: I4b7a4d251cdb50d5fa32fccb50910c9decc730d9
Bug-Url: https://bugzilla.redhat.com/1391130
Signed-off-by: Dominik Holler <dholler@redhat.com>
(cherry picked from commit 7a38c09d9a0f1691327f6620b2e198e0cd7dd651)