core: refactored overgrown method
as command itself is named: Activate[or]Deactivate … ~ this is
overgrown. Method executeVmCommand is repeatedly asking over and over,
what it actually should do, plug or unplug. So it's quite clear, that
this method should be split, which is very easy to do.

work process:
1. implement new method with switch.
2. duplicate original method, one for plug, one for unplug.
3. since swith already made decision, we might re-set variables
isNicToBePlugged and isNicToBeUnplugged to constants, true/false.
4. fix all static analysis problems. Example "if (null != null) {}"
can be removed.

both method should be much easier to grasp, since both methods now
fit PC screen, lost of conditions simply disappeared, and we are
solving one usecase at time. Unplug method became very short and easy.

Change-Id: Ifbce80c40105e1ac41529752eac099125a9ca1bc
Signed-off-by: Martin Mucha <mmucha@redhat.com>