core: fix the upgrade script that converts memory volumes to disks
This patch fixes yet another problem that was found in the upgrade
script that converts snapshot's memory volumes to disks in the
database. This time the problem is with using 'NOT IN <set>' where
the set could have NULL values (therefore the fix is to filter the
NULL values from the set).

Change-Id: I48103802395e58afb155a4501e36c256dbf65084
Bug-Url: https://bugzilla.redhat.com/1390254
Signed-off-by: Arik Hadas <ahadas@redhat.com>