core: Rename upload commands and variables to "transfer"
Since the current image upload commands are really about managing the
image transfer, and not specifically "uploading", its name should
reflect that. This can be acheived by renaming the command to
"transfer" commands, E.g:

UploadDiskImageCommand -> TransferDiskImageCommand

This will also allow the addition of the "direction" attribute for a
transfer, which states whether the transfer is an upload or a download,
supporting a download functionallity.

Change-Id: I398e487829d3773a4dd507c9fead4d130c6e798e
Bug-Url: https://bugzilla.redhat.com/1314387
Signed-off-by: Amit Aviram <aaviram@redhat.com>