core: rm ConfigureConsoleOptionsQuery#getConfig
Inline the useless getConfig method and use the standard MockConfigRule
supplied by the AbstractQueryTest#mcr member to mock config values where
needed.

Note that MockConfigRules mocks the default values for any ConfigValues
that is not supplied, so a lot of the boiler-plate mockings were just
removed for brevity's sake (see commits 9395f and 5b3463 for details).

Change-Id: If94b03439405a9154b120784a1256ad934e9f477
Signed-off-by: Allon Mureinik <amureini@redhat.com>