core: introduce Pass Discard support for vm disks
This patch adds the backend support for a new vm disk property - "Pass
Discard".

Passing discard can be set for a vm disk image on file and block storage
domains or for a vm direct lun, with an IDE or VirtIO-SCSI interface.
When Pass discard is enabled, if a discard command (UNMAP SCSI command)
is sent from the guest, the vm will not ignore it and will pass it on to
the underlying storage.

By enabling the Pass Discard property for a vm disk, thin provisioned
luns in the underlying storage that are used as direct luns or that are
used to create a block storage domain can not only grow, but also can
shrink. With that, unused space can be "returned" to the storage
(recycled) and be used by someone else.

Change-Id: Iec84aeddbe7f088e989b4d9dbaadec9364671971
Bug-Url: https://bugzilla.redhat.com/1241106
Signed-off-by: Idan Shaby <ishaby@redhat.com>