core: remove redundant comparison in vm-analyzer
When we detect a VM that is running on a host that we
don't expect it to run on (not as part of migration),
we should update the VM so its runOnVds will point to
the new host. There is no point in comparing the VM's
runtime information we received from VDSM with the
information we have in the database.

Change-Id: I719571e35c08125eb73ff7dadcf85a4e9b5f7ba3
Signed-off-by: Arik Hadas <ahadas@redhat.com>