core: prevent NPE on unmanaged migrations
During migrations that were not initiated by the engine we
do not have the destination host, unfortunately. This patch
prevents the NPE that is then thrown in the hand-over by a
simple null-check. The VM would then be set to MigratingTo
and its runOnVds would be null. Thus, we enhance the VM
analysing with a proper handling for this new state - when
the VM would be detected on the destination host, its
runOnVds will change to point to that host.

Change-Id: Ie11c0eb14318d8ab28f613cbf012879aba4f2ab9
Bug-Url: https://bugzilla.redhat.com/1392903
Signed-off-by: Arik Hadas <ahadas@redhat.com>