core: remove incorrect logic in vm-analyzer
When we detect a VM that is running on a host that we did not
expect the VM to run on, we first examine the migration
statuses MigratingFrom and MigratingTo. Then, we used to have
the following logic: "if the VM is set as running on a
different host and it is not up on this host then it is being
migrated". This logic is obviously wrong - except for Down
state that is handled before, all other statuses should be
treated as the Up status.

Change-Id: Ib9cccca87b251810c6260ca14807e4c2d66d5b16
Signed-off-by: Arik Hadas <ahadas@redhat.com>