packaging: setup: more severe el6 hosts check
Due to other changes, upgrading the engine to 3.6
with hosts still running hosted-engine from 3.5 on el6
could be by far more problematic than in the past.
In the past we were advising but the user could eventually
ignore and continue from the interactive setup.
Now the user can only force to skip that check via answerfile;
the interacive upgrade will always abort if we have el6
based hosted-engine hosts.

Change-Id: Ib78a6347f458d715f79f0247de5f3c744c810813
Bug-Url: https://bugzilla.redhat.com/1395357
Signed-off-by: Simone Tiraboschi <stirabos@redhat.com>