core: Move initialization in LiveMigrateDiskCommand to init block
LiveMigrateDiskCommand CTOR is calling getVm() which triggers a DAO call,
those calls in the command classes are now using the DAOs by injection and
since injection is performed after the CTOR is invoked the DAOs will be
null.
Moved the initialization part from the CTOR to the init block where it
belonds to avoid the NPE caused.

Change-Id: Idbc71d59f8c2f47e320bed2b98056fa102fa84db
Bug-Url: https://bugzilla.redhat.com/1395746
Signed-off-by: Tal Nisan <tnisan@redhat.com>