core: enable entity progress support for all steps
Currently entity progress report is supported for commands that
add a step by overriding getCommandStep() and getCommandStepSubjectEntities().

On flows adding steps using different infrastructures (like async task manager
based flows) it may be required to report progress on the entities as well.

The progress on the entity is a sum of the progress*entity weight
for all the steps "related" to the entity (we currently use that on jobs
that modifies an entity exclusively). Therefore In order to support
entity progress report for all the steps, we need a progress for each of
them and to support for adding the step subject entities.

This patch includes the following changes:
1. Add support to include StepSubjectEntity list when adding any step.
2. Move the StepSubjectEntity handling to the JobRepository
in order to reuse the code by all the callers.
3. Adding the step_progress view so that steps without progress report
will be considered with progress of 100 when they are finished.
4. Update to the entity_step_progress accordingly.

Change-Id: I1511291b18dbbb6e11f715a9a1e203c839cbaffa
Signed-off-by: Liron Aravot <laravot@redhat.com>