core: add ID property to MacPool.
Adding ID to MacPool to be able to identify different MacPools in the
logs. The DB ID (mac_pools.id) will be used for this.

Note:
ID is not a 'required' property of in-memory object (with an exception
in compensation flow), and might bind itself to DB, it *is* a property
of MacPool, and a good candidate to identify pool. A pool is related
to db record, and has ID. It's very unlikely, that any future
implementation of pools won't be able to have ID, thus proceeding
without pool having ID is just complicating work for ourselves.

Change-Id: Icb688c24e9105850e70291a67d7fcc129773a9ca
Bug-Url: https://bugzilla.redhat.com/1395462
Signed-off-by: Martin Mucha <mmucha@redhat.com>