engine: Report upgrade packages list only when relevant
There are cases in which the reported list of packages available for
updates is differ than the queried list of packages.
Since the user is interested the preconfigured packages for query, the
reported list should include the list of packages only when available
and log a general message without the packages otherwise.

Change-Id: I54f3bcd68abc0d9356f0e330a10978d629c7662c
Bug-Url: https://bugzilla.redhat.com/1381419
Signed-off-by: Moti Asayag <masayag@redhat.com>