restapi: Add ticket action to vm graphics console
Previously it was only possible to request console ticket at the VM level which
in cases of SPICE+VNC automatically preferred SPICE thereby preventing the user
from requesting a VNC ticket.

Now also the per-graphics-console
/vms/{vm_id}/graphicsconsoles/{console_id}/ticket action is supported when user
wants to explicitly decide which console to use.

The logic in BackendVmResource was moved to BackendVmGraphicsConsoleResource
and the vm resource now (based on the same logic as before) delegates to the
appropriate graphics console subresource.

Change-Id: I5999171cf4f9d0609f258d329efd1240974aa815
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1383342
Signed-off-by: Martin Betak <mbetak@redhat.com>