webadmin: Fix vNIC profile page for ja_JP locale
Because the Japanese word for the checkbox label "public" consumes more
horizontal space than then the English word, the VnicProfileWidget
looked distorted a bit as the help button appears on a different line
below other items on the page.

Since the available horizontal space was allocated already, the margin
of the label is squeezed, to avoid a second row for the
VnicProfileWidget.

Change-Id: I2b423d1d8353d7f72f119f22763a1034d324ab1e
Bug-Url: https://bugzilla.redhat.com/1255257
Signed-off-by: Dominik Holler <dholler@redhat.com>