core: Improve comment phrasing: insure->ensure
While US spelling allows "insure" as an alternative spelling to
"ensure", this spelling is mainly associated with providing
compensation if a certain risk occurs. "Ensure", on the other hand, is
a commonly accepted spelling for making a guarantee.

This patch improves the spelling in an inline comment to reduce any
ambiguities that sentence may have had, and fixes the capitalization
and grammar issues it has.

Change-Id: Ia1fcf19ff691af2dcf499c249c45a79f4543f8ee
Signed-off-by: Allon Mureinik <amureini@redhat.com>