frontend: Properly enable instance images on New VM from Blank tmpl.
In the case the user switches to blank template the initDisks() method which
among other things also updates the availability of instance images editor is
not called therefore after switching from template with 1+ disks the editor
stayed disabled even for the blank template.

Fixed by addding enablement of the instance images editor also to the specific
flow for the blank template in the NewVmModelBehavior.

Change-Id: I3cd612ac0a961a79e9a87ca46e28f6c877388d71
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1376339
Signed-off-by: Martin Betak <mbetak@redhat.com>