vdsbroker: mark vds as spm candidate before connectStoragePool
When we assign the spm role we first perform connectStoragePool()
to make sure that the host has the latest pool metadata.
If connectStoragePool fails on the master domain being unavailable
we'll attempt to perform failover (start the spm on a different host)
if there are more hosts in the data center. When there are no more
candidate hosts for the spm role reconstruct will be performed.

Because the failure is on the connect operation, the host won't be
marked as the current candidate for the spm rule which will prevent
failover from running, causing for reconstruct to not be executed.

Change-Id: I65a44d9c42782ea27aa257e5b53ccb70c9dbcca8
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1396183
Signed-off-by: Liron Aravot <laravot@redhat.com>