webadmin: Fix selection of placeholder host devices in VmHostDevices subtab
It wasn't possible to unassign VM host devices when the selection contained
IOMMU placeholder devices (which on their own cannot be removed).

Now we consider as "selected" only the primary (non-placeholder) host devices
and enable/disable the "Remove" button accordingly. Also on the backend are
sent only those "primary" devices as the current contract of
Add/RemoveVmHostDevices command expects that.

Change-Id: I33537c4039eee9c453ca330059420ae038129820
Bug-Url: https://bugzilla.redhat.com/1343870
Signed-off-by: Martin Betak <mbetak@redhat.com>