core: Align 'httpcomponents' version with WildFly
The different components of the 'httpcomponents'
project are released indepedently. In particular
the 'core' and 'client' parts are released
independently and they have different version
numbers. In particular the versions included in
WildFly are 'httpcore-4.4.1' and 'httpclient-4.5'.
But in our POM files we have a single
'httpcomponents.version' property and we assume
that they have all the same version. To align
with what is really used during runtime this
patch removes that property and add the explicit
version numbers to for each component.

Change-Id: I9211e09b81af582eeb3b6a04fe4d526c699f0394
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>