core: Enable disk transferring for users via REST.
This patch enables image transferring for not-admin users.

Until now, it was prevented in the system due to two main reasons:

1. GetImageTransferByIdQuery was an Admin query, and is being called
from image transfer REST commands. This patch changes it to a User query
and makes the needed changes in the transfer DAO in terms of filtering,
etc.

2. Transfer commands did not have any action group, making it
unavailable for users with permissions to transfer an image. This patch
defines those commands to belong to EDIT_DISK_PROPERTIE action group-
which is the needed one to do a transfer.

Change-Id: Icf8311d9d53ab88f7e3305fb8fa0ed4cd4ff078c
Signed-off-by: Amit Aviram <aaviram@redhat.com>
Signed-off-by: Daniel Erez <derez@redhat.com>