core: reconstruct never succeeds when the master is v4
Currently we use the DisconnectPoolOnReconstruct config value to
determine whether its needed to disconnect from the storage pool
or not during reconstruct.
When V4 was added, it wasn't added to this config value. While we
should change the disconnect to be true by default - this patch adds it
to the config value in the meanwhile.

Change-Id: I64e24cf7f31926415170085a8ca79696796c93e6
Signed-off-by: Liron Aravot <laravot@redhat.com>