Add support in engine for Headless VM
This patch add support in engine for Headless VMs.
It is done in UI by adding a checkbox for enable/disable it
and is implemented by removing all graphical consoles and VIDEO devices
(aka Display type) for the VM before starting it.
The serial console is handled as before.

The following entities are fully handled in this patch (new, edit, create from
template, create template from vm) for both webadmin and userportal:
VM dialog
pool dialog
template dialog
instance type
Run Once (headless VM can run as non-headless in 1st run and
non-headless can run as headless)

The following are not part of this patch and will be added in following
patches:
-REST api support - backward compatability is fully supported but this
new headless feature is still not

Change-Id: Idea9d5e57b9bc06b8c0e74a8e627020c75116045
Bug-Url: https://bugzilla.redhat.com/1381184
Signed-off-by: Sharon Gratch <sgratch@redhat.com>