engine: improve handling of 3.6 hosts in 4.x clusters
Hosts not supporting cluster versions >= 4.0 does not report their
switch type. Such hosts to a cluster versions >= 4.0 get in a blocked
state, because their capabilities could not parsed because of the
missing switch type.

This patch makes the switch type required only for hosts, which claim
to support a cluster version >= 4.0, independently of the used
compatibility version of the cluster. This enables hosts to fail nicely by the
HandleVdsVersionCommand when added to a cluster of an unsupported
version.

Change-Id: I6e44f2650daf509216435e2d8f92310486cd9864
Bug-Url: https://bugzilla.redhat.com/1375573
Signed-off-by: Dominik Holler <dholler@redhat.com>