core: Remove ChangeFloppy functionality
The ChangeFloppy functionality is not used anywhere in the code, and
could (should) therefore be removed to make the codebase easier to
maintain.

This claim is confirmed by the fact the ChangeFloppyCommand's
constructor calls getVm(), which, in turn calls VmDao. Commit
f0dc3d (merged on 25/10/2016) introduced injection for this Dao in
AuditLoggableBase, meaning that the DAO would not yet be initialized in
the constructor, and calling it would fail with a NullPointerException,
making this command unusable.

This patch removes the following:
1. The ChangeFloppyCommand and its corresponding enum constant in
   VdcActionType.
2. The ChangeFloppyVDSCommand it calls, its corresponding enum
   constant in VDSCommandType, the IVdsServer method it calls and its
   implementation in JsonRpcVdsServer.
3. AuditLog constants and messages only used by this command.

Change-Id: I2ef2972e120a357bd13511b502ce0db8b35711fe
Signed-off-by: Allon Mureinik <amureini@redhat.com>