userportal, webadmin: tooltips performance improvement
Tooltips performance improvements for 3.6. We're removing
almost all uses of Bootstrap tooltips (used via gwtbootstrap3
library) because we've determined that they are leaking
memory and causing severe performance problems with the
amount of them we were using.

As an alternative, we're falling back to using the default
browser-based tooltips by setting the 'title' attribute on
elements.

A downside to this approach is the loss of HTML functionality
in the tooltips. In a few rare places where we need very
expressive tooltips, like the 'Setup Networks' dialog, we've
left the use of the Bootstrap tooltips.

Change-Id: I40436cff6951d26331a91ad8db5ac7dbd6ab90f9
Bug-Url: https://bugzilla.redhat.com/1399489
Signed-off-by: Greg Sheremeta <gshereme@redhat.com>