engine: Reduce the severity of check-upgrade log
When there is an attempt to check for an upgrade, while another process
is already running, the reported message is logged with an ERROR
severity. This specific message severity can be reduced to warning.

In addition, there is an inflation of error messages during the
process which can be reduced.

Change-Id: Ia325f85527491ce2c1a3a16263d2b68867f63b2a
Bug-Url: https://bugzilla.redhat.com/1400500
Signed-off-by: Moti Asayag <masayag@redhat.com>